# Conflicts:
#	composer.json

+++ protected/humhub/modules/activity/Events.php
@@ -12,10 +12,11 @@ use Yii;
 use humhub\modules\activity\components\MailSummary;
 use humhub\modules\activity\jobs\SendMailSummary;
 use humhub\modules\activity\models\Activity;
+use yii\base\Event;
 
 /**
  * Events provides callbacks to handle events.
- * 
+ *
  * @author luke
  */
 class Events extends \yii\base\Object
@@ -23,7 +24,7 @@ class Events extends \yii\base\Object
 
     /**
      * Handles cron run event to send mail summaries to the users
-     * 
+     *
      * @param \yii\base\ActionEvent $event
      */
     public static function onCronRun($event)
@@ -32,20 +33,34 @@ class Events extends \yii\base\Object
             Yii::$app->queue->push(new SendMailSummary(['interval' => MailSummary::INTERVAL_HOURY]));
         } elseif (Yii::$app->controller->action->id == 'daily') {
             Yii::$app->queue->push(new SendMailSummary(['interval' => MailSummary::INTERVAL_DAILY]));
+            if (date('N') == Yii::$app->getModule('activity')->weeklySummaryDay) {
+                Yii::$app->queue->push(new SendMailSummary(['interval' => MailSummary::INTERVAL_WEEKLY]));
+            }
         }
     }
 
     /**
      * On delete of some active record, check if there are related activities and delete them.
+     *
+     * @param Event $event
      */
-    public static function onActiveRecordDelete($event)
+    public static function onActiveRecordDelete(Event $event)
     {
-        $model = $event->sender->className();
-        $pk = $event->sender->getPrimaryKey();
+        if (!($event->sender instanceof \yii\db\ActiveRecord)) {
+            throw new \LogicException('The handler can be applied only to the \yii\db\ActiveRecord.');
+        }
+
+        /** @var \yii\db\ActiveRecord $activeRecordModel */
+        $activeRecordModel = $event->sender;
+        $pk = $activeRecordModel->getPrimaryKey();
 
         // Check if primary key exists and is not array (multiple pk)
         if ($pk !== null && !is_array($pk)) {
-            foreach (models\Activity::find()->where(['object_id' => $pk, 'object_model' => $model])->all() as $activity) {
+            $modelsActivity = Activity::find()->where([
+                'object_id' => $pk,
+                'object_model' => $activeRecordModel::className(),
+            ])->each();
+            foreach ($modelsActivity as $activity) {
                 $activity->delete();
             }
         }
